+++ app/_core/components/Pagination/Pagination.container.tsx
@@ -1,5 +1,5 @@
 import { DEFAULT_PAGE_SIZE } from "@/app/(protected)/(posts)/(modules)/categories/consts/pagination"
-import { PaginationControls } from "@/app/_core/components/PaginationControls"
+import { PaginationControls } from "@/app/_core/components/Pagination/PaginationControls"
 import { SettingKey } from "@/app/(protected)/(posts)/(modules)/settings/lib/settings.interface"
 import { settingsService } from "@/app/(protected)/(posts)/(modules)/settings/lib/settings.service"
