+++ app/_core/components/Pagination/pagination.texts.ts
@@ -1,0 +1,4 @@
+export const TEXTS = {
+    nextPage: "עמוד הבא",
+    previousPage: "עמוד קודם",
+}
